requirements.txt
tk
loguru
requests
pillow